ugc_service/ugc/requirements.txt
fastapi==0.100.1
uvicorn==0.12.2
aiokafka==0.10.0
orjson==3.9.15
pydantic-settings==2.2.1
